Move getFileManagerLabel into its own file-manager module
