+++ apps/studio/src/components/site-menu.tsx
@@ -9,8 +9,9 @@ import { useContentTabs } from 'src/hooks/use-content-tabs';
 import { useDeleteSite } from 'src/hooks/use-delete-site';
 import { useImportExport } from 'src/hooks/use-import-export';
 import { useSiteDetails } from 'src/hooks/use-site-details';
-import { getFileManagerLabel, isMac } from 'src/lib/app-globals';
+import { isMac } from 'src/lib/app-globals';
 import { cx } from 'src/lib/cx';
+import { getFileManagerLabel } from 'src/lib/file-manager';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { supportedEditorConfig } from 'src/modules/user-settings/lib/editor';
 import { getTerminalName } from 'src/modules/user-settings/lib/terminal';
