+++ apps/studio/src/components/tests/content-tab-overview.test.tsx
@@ -12,8 +12,10 @@ import { testActions, testReducer } from 'src/stores/tests/utils/test-reducer';
 vi.mock( 'src/lib/app-globals', () => ( {
 	isWindows: vi.fn().mockReturnValue( false ),
 	isLinux: vi.fn().mockReturnValue( false ),
-	getFileManagerLabel: vi.fn().mockReturnValue( 'Finder' ),
 	getAppGlobals: vi.fn( () => ( { platform: 'darwin' } ) ),
+} ) );
+vi.mock( 'src/lib/file-manager', () => ( {
+	getFileManagerLabel: vi.fn().mockReturnValue( 'Finder' ),
 } ) );
 vi.mock( 'src/hooks/use-site-details' );
 vi.mock( 'src/hooks/use-theme-details' );
