+++ apps/studio/src/lib/app-globals.ts
@@ -1,5 +1,3 @@
-import { __ } from '@wordpress/i18n';
-
 export function getAppGlobals(): AppGlobals {
 	return window.appGlobals;
 }
@@ -27,17 +25,4 @@ export function isLinux() {
 
 export function isWindowsStore() {
 	return getAppGlobals().isWindowsStore;
-}
-
-export function getFileManagerLabel(): string {
-	if ( isWindows() ) {
-		// translators: name of app used to navigate files and folders on Windows
-		return __( 'File Explorer' );
-	}
-	if ( isLinux() ) {
-		// translators: generic name of the app used to navigate files and folders on Linux
-		return __( 'File Manager' );
-	}
-	// translators: name of app used to navigate files and folders on macOS
-	return __( 'Finder' );
 }
